requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
starlette==0.52.1
httpx==0.28.1
asyncpg==0.31.0
pydantic==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.22
